Post Entity has been done!

+++ src/main/java/com/photogram/entity/Comment.java
@@ -5,9 +5,9 @@ import lombok.Data;
 import java.sql.Timestamp;
 @Data
 public class Comment {
-    private int id;
-    private int postId;
-    private int userId;
+    private Long id;
+    private Long postId;
+    private Long userId;
     private String text;
     private Timestamp commentTime;
 }
