+++ src/main/java/com/photogram/entity/Image.java
@@ -4,9 +4,9 @@ import lombok.Data;
 
 @Data
 public class Image {
-    private int id;
+    private Long id;
     private String path;
-    private int postId;
-    private int userId;
+    private Long postId;
+    private Long userId;
 
 }
