+++ src/main/java/com/photogram/entity/PostLike.java
@@ -6,8 +6,8 @@ import lombok.Data;
 import java.sql.Timestamp;
 @Data
 public class PostLike {
-    private int id;
-    private int userId;
-    private int postId;
+    private Long id;
+    private Long userId;
+    private Long postId;
     private Timestamp likeTime;
 }
